Correções diversas.

+++ ADA.Consumer/DTO/TransacaoDTO.cs
@@ -8,7 +8,7 @@ public class TransacaoDTO
     public string ContaOrigem { get; set; } = null!;
     public string ContaDestino { get; set; } = null!;
     public Canal Canal { get; set; }
-    public float Valor { get; set; }
+    public double Valor { get; set; }
     public double Latitute { get; set; }
     public double Longitude { get; set; }
